OS: FCFS and RR Implemented

+++ 3rd year/2nd semester/CSE 322 Operating Systems/OSScheduler/OSScheduler/GUI/TimeMarker.xaml.cs
@@ -13,7 +13,7 @@ namespace OSScheduler.GUI
             
         }
 
-        public TimeMarker(int time) : this()
+        public TimeMarker(double time) : this()
         {
             Time.Content = time;
         }
